My Update
1. Added the services, resource, projects, and contact us tabs (jsp &
controller).  The controllers for those tabs are pretty much all the
same.  They are a copy of the LoginController.

2. Still working on registration page.  I include some front end error
checking. The registration page commits to the database but I still need
to have it pass in the role and discipline (those are hard coded to
CompSci and student for now).