Registration completed
1. I deleted Mike's EmailValidator class because it is not needed. I
tested for valid email addresses using a regular expression I got from:
http://www.mkyong.com/regular-expressions/how-to-validate-email-address-with-regular-expression/

2. I added a check to test the length of the email address to be <=40
characters. The database is currently set to store up to 80 characters
but I will change this to 45. While researching the average email
lengths in a database, it turns out addresses fall off sharply at 38
characters.

3. I changed the name validation from 4-20 characters to <=40
characters. There are some names that are less than 4 characters. (i.e.
A first name Ed or a last name Ng.). Also, the max field in the database
is 45.

4. On the registration form, I put line breaks after the field titles so
that the errors fall right below the titles.  I think this looks
cleaner.

Registration is completed.