Profile Page.
The profile successfully writes into the database.  I created a generic
profilepage.jsp because the other one was giving me problems.

I changed some other classes as well.  One change in particular was in
the LoginController - I changed the role if-then to a case/switch
statement.

Please look it over.